requirements.txt
# Cloud FinOps Toolkit - Python Dependencies

# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3

# Cloud Provider SDKs
boto3==1.34.20
google-cloud-bigquery==3.14.1
azure-mgmt-costmanagement==4.0.1
azure-mgmt-consumption==10.0.0
azure-identity==1.15.0

# Data Processing
pandas==2.1.4
numpy==1.26.3
duckdb==0.9.2

# Anomaly Detection
ruptures==1.1.9
prophet==1.1.5
scikit-learn==1.4.0

# Database & Storage
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
redis==5.0.1

# Monitoring & Logging
prometheus-client==0.19.0
python-json-logger==2.0.7

# OPA Integration (optional)
# Note: OPA itself should be installed separately

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0

# Utilities
python-dotenv==1.0.0
click==8.1.7
pyyaml==6.0.1
python-dateutil==2.8.2

# Slack/Jira Integration (optional)
slack-sdk==3.26.2
jira==3.6.0

# Development Tools
black==24.1.1
flake8==7.0.0
mypy==1.8.0
pre-commit==3.6.0
